requirements.txt
Flask==2.2.2
google-cloud-datastore==2.7.1
json2html==1.3.0
google-cloud-secret-manager==2.12.6
google-auth-oauthlib==0.7.1